LayoutTests/imported/w3c/web-platform-tests/svg/geometry/svg-baseval-in-display-none-expected.txt


PASS With em
FAIL With em and percentage assert_equals: r1.height expected 120 but got 120.00000762939453
